+++ src/main/kotlin/com/kweather/domain/weather/controller/WeatherController.kt
@@ -1,39 +1,42 @@
 package com.kweather.domain.weather.controller
 
 import com.kweather.domain.airstagnation.dto.AirStagnationIndexInfo
+import com.kweather.domain.forecast.dto.ForecastInfo
+import com.kweather.domain.realtime.dto.RealTimeDustInfo
+import com.kweather.domain.senta.dto.SenTaIndexInfo
+import com.kweather.domain.uvi.dto.UVIndexInfo
 import com.kweather.domain.weather.entity.Weather
 import com.kweather.domain.weather.model.AirQuality
 import com.kweather.domain.weather.model.HourlyForecast
 import com.kweather.domain.weather.model.UVIndex
-import com.kweather.domain.weather.service.WeatherService
+import com.kweather.domain.weather.model.WeatherDataProvider
+import com.kweather.domain.weather.service.GeneralWeatherService
 import com.kweather.global.common.util.DateTimeUtils
+import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
 import java.time.LocalDateTime
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
-import org.slf4j.LoggerFactory
-import com.kweather.domain.weather.model.WeatherDataProvider
-import com.kweather.domain.realtime.dto.RealTimeDustInfo
-import com.kweather.domain.senta.dto.SenTaIndexInfo
-import com.kweather.domain.uvi.dto.UVIndexInfo
-import org.springframework.beans.factory.annotation.Value
 import arrow.core.Option
 import arrow.core.Some
 import arrow.core.None
+import com.kweather.domain.airstagnation.service.AirStagnationIndexService
+import com.kweather.domain.senta.service.SenTaIndexService
+import com.kweather.domain.uvi.service.UVIndexService
 
 @Controller
 class WeatherController(
-    private val weatherService: WeatherService,
-    @Value("\${weather.default.region.name:한남동}")
-    private val defaultRegionName: String,
-    @Value("\${weather.default.region.district:용산구}")
-    private val defaultRegionDistrict: String,
-    @Value("\${weather.default.sido:서울}")
-    private val defaultSido: String,
-    @Value("\${weather.default.area-no:1100000000}")
-    private val defaultAreaNo: String
+    private val generalWeatherService: GeneralWeatherService,
+    private val uvIndexService: UVIndexService,
+    private val senTaIndexService: SenTaIndexService,
+    private val airStagnationIndexService: AirStagnationIndexService,
+    @Value("\${weather.default.region.name:한남동}") private val defaultRegionName: String,
+    @Value("\${weather.default.region.district:용산구}") private val defaultRegionDistrict: String,
+    @Value("\${weather.default.sido:서울}") private val defaultSido: String,
+    @Value("\${weather.default.area-no:1100000000}") private val defaultAreaNo: String
 ) {
     private val logger = LoggerFactory.getLogger(WeatherController::class.java)
 
@@ -41,11 +44,11 @@ class WeatherController(
         override fun getWeatherData(date: String, time: String): Weather =
             createDefaultWeatherData(date, time)
 
-        override fun getDustForecastData(date: String, informCode: String): List<com.kweather.domain.forecast.dto.ForecastInfo> =
+        override fun getDustForecastData(date: String, informCode: String): List<ForecastInfo> =
             runCatching {
-                weatherService.getDustForecast(date, informCode)
+                generalWeatherService.getDustForecast(date, informCode)
             }.getOrElse { e ->
-                logger.error("Failed to get dust forecast: ${e.message}", e)
+                logger.error("미세먼지 예보 데이터 가져오기 실패: ${e.message}", e)
                 emptyList()
             }
 
@@ -65,33 +68,33 @@ class WeatherController(
     private inner class LiveWeatherDataProvider : WeatherDataProvider {
         override fun getWeatherData(date: String, time: String): Weather =
             runCatching {
-                weatherService.buildWeatherEntity(60, 127)
+                generalWeatherService.buildWeatherEntity(60, 127)
             }.getOrElse { e ->
-                logger.error("Failed to get live weather data: ${e.message}", e)
+                logger.error("실시간 날씨 데이터 가져오기 실패: ${e.message}", e)
                 createDefaultWeatherData(date, time)
             }
 
-        override fun getDustForecastData(date: String, informCode: String): List<com.kweather.domain.forecast.dto.ForecastInfo> =
+        override fun getDustForecastData(date: String, informCode: String): List<ForecastInfo> =
             runCatching {
-                weatherService.getDustForecast(date, informCode)
+                generalWeatherService.getDustForecast(date, informCode)
             }.getOrElse { e ->
-                logger.error("Failed to get dust forecast: ${e.message}", e)
+                logger.error("미세먼지 예보 데이터 가져오기 실패: ${e.message}", e)
                 emptyList()
             }
 
         override fun getRealTimeDustData(sidoName: String): List<RealTimeDustInfo> =
             runCatching {
-                weatherService.getRealTimeDust(sidoName)
+                generalWeatherService.getRealTimeDust(sidoName)
             }.getOrElse { e ->
-                logger.error("Failed to get real-time dust data: ${e.message}", e)
+                logger.error("실시간 미세먼지 데이터 가져오기 실패: ${e.message}", e)
                 emptyList()
             }
 
         override fun getUVIndexData(areaNo: String, time: String): List<UVIndexInfo> =
             runCatching {
-                weatherService.getUVIndex(areaNo, time)
+                uvIndexService.getUVIndex(areaNo, time)
             }.getOrElse { e ->
-                logger.error("Failed to get UV index data: ${e.message}", e)
+                logger.error("자외선 지수 데이터 가져오기 실패: ${e.message}", e)
                 emptyList()
             }
 
@@ -99,20 +102,20 @@ class WeatherController(
             runCatching {
                 val currentMonth = LocalDateTime.now(ZoneId.of("Asia/Seoul")).monthValue
                 if (currentMonth in 5..9) {
-                    weatherService.getSenTaIndex(areaNo, time)
+                    senTaIndexService.getSenTaIndex(areaNo, time)
                 } else {
                     emptyList()
                 }
             }.getOrElse { e ->
-                logger.error("Failed to get sensible temperature index data: ${e.message}", e)
+                logger.error("여름철 체감온도 데이터 가져오기 실패: ${e.message}", e)
                 emptyList()
             }
 
         override fun getAirStagnationIndexData(areaNo: String, time: String): List<AirStagnationIndexInfo> =
             runCatching {
-                weatherService.getAirStagnationIndex(areaNo, time)
+                airStagnationIndexService.getAirStagnationIndex(areaNo, time)
             }.getOrElse { e ->
-                logger.error("Failed to get air stagnation index data: ${e.message}", e)
+                logger.error("대기정체지수 데이터 가져오기 실패: ${e.message}", e)
                 emptyList()
             }
     }
@@ -182,7 +185,7 @@ class WeatherController(
 
     private fun getCurrentDateTimeInfo(): Triple<String, String, Int> {
         val (date, time) = DateTimeUtils.getCurrentDateTimeFormatted()
-        val hour = DateTimeUtils.getCurrentHour() // 현재 시간: 21
+        val hour = DateTimeUtils.getCurrentHour()
         return Triple(date, time, hour)
     }
 
@@ -193,7 +196,7 @@ class WeatherController(
     private fun createWeatherDataProvider(useRealTimeData: Boolean): WeatherDataProvider =
         if (useRealTimeData) LiveWeatherDataProvider() else DefaultWeatherDataProvider()
 
-    private fun processDustForecastData(dustForecast: List<com.kweather.domain.forecast.dto.ForecastInfo>): List<Triple<List<String>, List<String>, List<String>>> =
+    private fun processDustForecastData(dustForecast: List<ForecastInfo>): List<Triple<List<String>, List<String>, List<String>>> =
         dustForecast.map { forecast ->
             val regions = parseRegionGrades(forecast.grade)
             Triple(
@@ -225,7 +228,7 @@ class WeatherController(
     private fun addAttributesToModel(
         model: Model,
         weatherData: Weather,
-        dustForecast: List<com.kweather.domain.forecast.dto.ForecastInfo>,
+        dustForecast: List<ForecastInfo>,
         realTimeDust: List<RealTimeDustInfo>,
         uvIndexData: List<UVIndexInfo>,
         senTaIndexData: List<SenTaIndexInfo>,
@@ -286,15 +289,15 @@ class WeatherController(
         )
 
     private fun createDefaultHourlyForecast(): List<HourlyForecast> {
-        val currentHour = DateTimeUtils.getCurrentHour() // 21
-        val currentHourIndex = currentHour / 3 * 3 // 21
+        val currentHour = DateTimeUtils.getCurrentHour()
+        val currentHourIndex = currentHour / 3 * 3
         val hours = listOf(0, 3, 6, 9, 12, 15, 18, 21)
         return hours.mapIndexed { index, hourOffset ->
-            val forecastHour = (currentHourIndex + hourOffset) % 24 // 21, 0, 3, 6, 9, 12, 15, 18
+            val forecastHour = (currentHourIndex + hourOffset) % 24
             val forecastTime = if (hourOffset == 0) "지금" else "${forecastHour}시"
             val forecastIcon = if (forecastHour in 6..18) "sun" else "moon"
-            val forecastTemp = "-${index + 1}°C" // 기본값, 실제 데이터로 대체 필요
-            val forecastHumidity = "${(50 + index * 5)}%" // 기본값, 실제 데이터로 대체 필요
+            val forecastTemp = "-${index + 1}°C"
+            val forecastHumidity = "${(50 + index * 5)}%"
             HourlyForecast(forecastTime, forecastIcon, forecastTemp, forecastHumidity)
         }
     }
